Use BigIntegerField for quota fields

+++ qabel_provider/models.py
@@ -21,9 +21,9 @@ class Profile(models.Model, ExportModelOperationsMixin('profile')):
     plus_notification_mail = models.BooleanField(default=False)
     pro_notification_mail = models.BooleanField(default=False)
 
-    block_quota = models.IntegerField(verbose_name='block server quota (in bytes)', default=2 * 1024**3)
-    monthly_traffic_quota = models.IntegerField(verbose_name='block server traffic quota per month (in bytes)',
-                                                default=20 * 1024**3)
+    block_quota = models.BigIntegerField(verbose_name='block server quota (in bytes)', default=2 * 1024**3)
+    monthly_traffic_quota = models.BigIntegerField(verbose_name='block server traffic quota per month (in bytes)',
+                                                   default=20 * 1024**3)
 
     @property
     def is_confirmed(self):
